+++ src/app/core/auth/auth.service.ts
@@ -1,3 +1,4 @@
+import { StaticService } from 'src/app/config/static.service';
 import { Injectable } from '@angular/core';
 import { map } from 'rxjs/operators';
 import { BehaviorSubject, Observable, of } from 'rxjs';
@@ -9,6 +10,7 @@ import { ApiService } from 'src/app/core/http/api/api.service';
 import { ObjectPermission, Objects, Permissions, UserRoles } from 'src/app/core/models/permissions';
 import { TokenService } from 'src/app/core/services/token-check/token.service';
 import { LogService } from './../services/logger/log.service';
+import { AuthTokenStructure } from './auth.interfaces';
 
 /**
  * auth.service.ts
@@ -25,11 +27,14 @@ export class AuthService {
   public userAuthenticated$: Observable<UserData>;
   // exp-Date in form _x.xxxxxx
   private EXPIRES_FACTOR = 1000;
+  // Logout if Token is expired on Page Start
+  private logOutAutomatically: boolean;
 
   constructor(
     private apiService: ApiService,
     private tokenService: TokenService,
-    private logService: LogService
+    private logService: LogService,
+    private staticService: StaticService
   ) {
     this.userAuthenticated$ = this.user$.pipe(
       map((user: User) => {
@@ -39,6 +44,8 @@ export class AuthService {
         };
       })
     );
+
+    this.logOutAutomatically = this.staticService.getAuthBehaviour().autoLogout;
   }
 
   /**
@@ -48,21 +55,24 @@ export class AuthService {
    * @returns
    */
   public login(email: string, password: string): Observable<User | null> {
+    this.tokenService.removeOldToken(); //TO DELETE
+    this.tokenService.removeAccessToken();
+    this.tokenService.removeRefreshToken();
+
     return this.apiService.loginUser(email, password).pipe(
       map((serverResponse: AuthResponseData) => {
         let tmpUser: User = null;
 
         if (serverResponse.access_token) {
-          tmpUser = this.createUserFromToken(serverResponse.access_token);
-          if (tmpUser) {
-            this.tokenService.saveToken(serverResponse.access_token, tmpUser.tokenExpirationDate);
-           // this.logService.log('AuthService', 'login', tmpUser);
-            this.logService.log('AuthService', 'login:', tmpUser.name);
-          }
-        }
-        if (!tmpUser) {
-          // eslint-disable-next-line no-console
-          console.warn('Auth:', 'Es wurde kein Token gesendet');
+          const decodedToken = this.tokenService.getDecodedToken(serverResponse.access_token);
+          const expirationDate = new Date(decodedToken.exp * this.EXPIRES_FACTOR);
+
+          tmpUser = this.createUserFromToken(decodedToken);
+          this.logService.log('AuthService Login-Exp:', tmpUser.name, expirationDate);
+          this.tokenService.saveAccessToken(serverResponse.access_token);
+          this.tokenService.saveRefreshToken(serverResponse.refresh_token);
+        } else {
+          this.logService.warn('AuthService', 'NoToken from Server');
         }
 
         this.user$.next(tmpUser);
@@ -74,17 +84,23 @@ export class AuthService {
   /**
    * Auto-Login reading localStorage
    */
-  public autoLogin() {
+  public autoLogin(): void {
     // Userdaten im localStorage
     let tmpUser: User = null;
-    const token = this.tokenService.getToken();
+    this.tokenService.removeOldToken();
+    const token = this.tokenService.getAccessToken();
 
     if (token) {
-      tmpUser = this.createUserFromToken(token);
-      if (tmpUser) {
-        this.logService.log('AuthService', 'autologin: ', tmpUser);
+      const decodedToken = this.tokenService.getDecodedToken(token);
+      const expirationDate = new Date(decodedToken.exp * this.EXPIRES_FACTOR);
+
+      if (this.tokenCheckOk(expirationDate)) {
+        tmpUser = this.createUserFromToken(decodedToken);
+        this.logService.log('AuthService', 'autologin: ', tmpUser.name);
       } else {
-        this.tokenService.removeToken();
+        this.logService.log('AuthService', 'autologout');
+        this.tokenService.removeAccessToken();
+        this.tokenService.removeRefreshToken();
       }
     }
     this.user$.next(tmpUser);
@@ -96,6 +112,40 @@ export class AuthService {
     return of(true);
   }
 
+  public updateUserSession(): Observable<User | null> {
+    const refreshToken = this.tokenService.getRefreshToken();
+    //console.log('Try to refresh token');
+    return this.apiService.updateUserSession(refreshToken).pipe(
+      map((serverResponse: AuthResponseData) => {
+        let tmpUser: User = null;
+
+        if (serverResponse.access_token) {
+          //console.table(serverResponse);
+          const decodedToken = this.tokenService.getDecodedToken(serverResponse.access_token);
+          const expirationDate = new Date(decodedToken.exp * this.EXPIRES_FACTOR);
+
+          tmpUser = this.createUserFromToken(decodedToken);
+          //this.logService.log('AuthService: Saved new Token', JSON.stringify(decodedToken));
+          this.logService.log(
+            'AuthService: Update-Token Login-Exp:',
+            tmpUser.name,
+            expirationDate
+          );
+          this.tokenService.saveAccessToken(serverResponse.access_token);
+          this.tokenService.saveRefreshToken(serverResponse.refresh_token);
+        } else {
+          this.logService.warn('AuthService', 'NoToken from Server');
+        }
+        this.user$.next(tmpUser);
+        return tmpUser;
+      })
+    );
+  }
+
+  /**
+   * not used at the moment
+   * @returns Observable<boolean>
+   */
   public logoutUserOnTokenExpired(): Observable<boolean> {
     this.logService.log('AuthService', 'logout(automatically)');
     this.signOff();
@@ -103,33 +153,56 @@ export class AuthService {
   }
 
   private signOff() {
-    this.tokenService.removeToken();
+    this.tokenService.removeAccessToken();
+    this.tokenService.removeRefreshToken();
     this.user$.next(null);
   }
 
-  private createUserFromToken(token: string): User {
+  /**
+   * not used at the moment
+   * @returns
+   */
+  private logoutWithServerLogout() {
+    this.apiService.logoutUser().subscribe({
+      next: (response: any) => {
+        this.logService.log('AuthService', 'logout(manually ok)', response);
+        this.signOff();
+      },
+      error: (error) => {
+        this.logService.log('AuthService', error);
+        this.signOff();
+      },
+    });
+    return of(true);
+  }
+
+  private createUserFromToken(decodedToken: AuthTokenStructure): User {
     let tmpUser: User = null;
-    const decoded = this.tokenService.getDecodedToken(token);
-    const expirationDate = new Date(decoded.exp * this.EXPIRES_FACTOR);
-    //console.log('EXP-DATE:', expirationDate);
+    // Featurepermission setzen
+    const user_role =
+      decodedToken.user_role == undefined ? UserRoles.DEFAULT : decodedToken.user_role;
+    const user_featurepermission = this.setUserPermissions(user_role);
+    // User erstellen
+    tmpUser = new User(
+      decodedToken.user_id,
+      undefined,
+      decodedToken.user_name,
+      user_role,
+      user_featurepermission
+    );
+    return tmpUser;
+  }
+
+  private tokenCheckOk(expirationDate: Date): boolean {
     const expirationDuration = expirationDate.getTime() - new Date().getTime();
 
     if (expirationDuration > 0) {
-      // Featurepermission setzen
-      const user_role = decoded.user_role == undefined ? UserRoles.DEFAULT : decoded.user_role;
-      const user_featurepermission = this.setUserPermissions(user_role);
-      // User erstellen
-      tmpUser = new User(
-        decoded.user_id,
-        undefined,
-        decoded.user_name,
-        user_role,
-        token,
-        expirationDate,
-        user_featurepermission
-      );
+      this.logService.log('AuthService:', 'TOKEN-EXP:', expirationDate);
+    } else {
+      this.logService.warn('AuthService:', 'Token EXPIRED', expirationDate);
     }
-    return tmpUser;
+
+    return expirationDuration > 0 || !this.logOutAutomatically;
   }
 
   private setUserPermissions(user_role: string | string[]): ObjectPermission[] {