+++ src/app/sites/offers/pages/offer-detail/offer-detail.component.ts
@@ -1,15 +1,10 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { NgbdModalLmsContentComponent } from 'src/app/sites/offers/components/modalWindows/modal-offer-to-lms/ngbd-modal-lmscontent';
-import { NgbdModalOfferDeleteComponent } from 'src/app/sites/offers/components/modalWindows/modal-offer-delete/ngbd-modal-offerdelete';
 import { OfferDataService } from 'src/app/core/data/offer/offer-data.service';
 import { Offer } from 'src/app/core/models/offer';
-import { User } from 'src/app/core/models/user';
 import { StaticService } from 'src/app/config/static.service';
 import { Objects, Permissions } from 'src/app/core/models/permissions';
-import { LogService } from 'src/app/core/services/logger/log.service';
 import { ErrorHandlerService } from 'src/app/core/services/error-handling/error-handling';
 
 @Component({
@@ -19,69 +14,64 @@ import { ErrorHandlerService } from 'src/app/core/services/error-handling/error-
 })
 export class OfferDetailComponent implements OnInit, OnDestroy {
   lnkOffers = this.staticConfig.getPathInfo().lnkOffers;
-  lnkAdminOfferEdit = this.staticConfig.getPathInfo().lnkAdminOfferEdit;
-  private userSub: Subscription;
-  private paramsSub: Subscription;
+  private userSub: Subscription | undefined;
+  private paramsSub: Subscription | undefined;
 
-  public offer: Offer;
-  public user: User;
-  public isLoading = false;
-  public subscribed = false;
-
+  offer: Offer | undefined;
   object = Objects;
   permission = Permissions;
 
+  isLoading: boolean;
   isError: boolean;
   errMessage: string;
 
   constructor(
     private offerDataService: OfferDataService,
     private route: ActivatedRoute,
-    private router: Router,
-    private modalService: NgbModal,
     private staticConfig: StaticService,
-    private logService: LogService,
     private errorHandler: ErrorHandlerService
-  ) {}
-
-  ngOnInit() {
-    this.paramsSub = this.route.paramMap.subscribe((params) => {
-      const offerId = +params.get('id');
-
-      this.isLoading = true;
-      this.errMessage = '';
-      this.isError = false;
-
-      this.userSub = this.offerDataService.getOfferDataForDetail(offerId).subscribe({
-        next: (data) => {
-          //this.logService.log('offer-detail', 'Detaildata', data);
-          this.user = data.user;
-          this.offer = data.offerData;
-          this.subscribed = false;
-          this.isLoading = false;
-          this.errMessage = '';
-          this.isError = false;
-        },
-        error: (error: Error) => {
-          this.isLoading = false;
-          this.isError = true;
-          //this.errMessage =this.errorHandler.getErrorMessage(error, 'offer');
-          // Besser immer die not-found-Fehlermeldung anzeigen.
-          this.errMessage = this.errorHandler.ERROR_MESSAGES.E404_OFFER_NOT_FOUND;
-        },
-      });
-    });
+  ) {
+    this.isError = false;
+    this.isLoading = false;
+    this.errMessage = '';
   }
 
-  deleteOffer() {
-    // TODO: Lädt-Anzeige an
-    this.offerDataService.deleteOffer(this.offer).subscribe({
-      next: (value) => {
-        this.router.navigate([this.lnkOffers]);
-      },
-      error: (error: Error) => {
-        // TODO: Rückmeldung?
-      },
+  ngOnInit() {
+    this.resetError();
+    this.isLoading = false;
+
+    // Daten laden, wenn sich der Übergabeparameter ändert
+    this.paramsSub = this.route.paramMap.subscribe((params) => {
+      const strParam = params.get('id');
+
+      if (strParam && strParam.length > 0) {
+        const offerId: number = +strParam;
+
+        if (Number.isNaN(offerId)) {
+          this.setError();
+          this.isLoading = false;
+
+        } else {
+          this.resetError();
+          this.isLoading = true;
+
+          this.userSub = this.offerDataService.getOfferDataForDetail(offerId).subscribe({
+            next: (data) => {
+              //this.logService.log('offer-detail', 'Detaildata', data);
+              this.offer = data.offerData;
+              this.resetError();
+              this.isLoading = false;
+            },
+            error: (error: Error) => {
+              this.setError();
+              this.isLoading = false;
+            },
+          });
+        }
+      } else {
+        this.setError();
+        this.isLoading = false;
+      }
     });
   }
 
@@ -91,62 +81,18 @@ export class OfferDetailComponent implements OnInit, OnDestroy {
    */
   subscribeToOffer() {}
 
-  forwardToLMS() {
-    if (this.offer.url == null || this.offer.url == undefined) {
-      this.logService.log('offer-detail:', 'PATH is not defined');
-      return;
-    }
-
-    if (!this.isExternalURL(this.offer.url)) {
-      this.logService.log('offer-detail: ', 'PATH is not valid: ', this.offer.url);
-      return;
-    } else {
-    }
-
-    if (window) {
-      (window as any).location.href = this.offer.url;
-      // Alternativ neues Fenster
-      //(window as any).open(this.offer.url, "_blank");
-    }
-  }
-
-  showModalWindowToLMS() {
-    const modalRef = this.modalService.open(NgbdModalLmsContentComponent, {
-      centered: true,
-    });
-    modalRef.componentInstance.title = this.offer.title;
-    modalRef.result.then(
-      (result) => {
-        // console.log('Weiter ', result);
-        this.forwardToLMS();
-      },
-      (reason) => {
-        // Cancel by button or ModalDismissReasons
-        // console.log('Cancel ', reason);
-      }
-    );
-  }
-
-  showModalWindowDeleteOffer() {
-    const modalRef = this.modalService.open(NgbdModalOfferDeleteComponent, {
-      centered: true,
-    });
-    modalRef.componentInstance.title = this.offer.title;
-    modalRef.result.then(
-      (result) => {
-        this.deleteOffer();
-      },
-      (reason) => {}
-    );
-  }
-
-  isExternalURL(str: string): boolean {
-    // Muss mit http(s):// anfangen
-    return /^http[s]?:\/\//.test(str);
-  }
-
   ngOnDestroy(): void {
     if (this.userSub) this.userSub.unsubscribe();
     if (this.paramsSub) this.paramsSub.unsubscribe();
+  }
+
+  private setError() {
+    this.isError = true;
+    this.errMessage = this.errorHandler.ERROR_MESSAGES.E404_OFFER_NOT_FOUND;
+  }
+
+  private resetError() {
+    this.errMessage = '';
+    this.isError = false;
   }
 }